REWRITE: custom evaluator, use var instead of gen-class instances as fixture objects
We no longer use the default OGNL Evaluator and thus do not need to provide a class with particular methods.
We use just a var (holding the options map) as the fixture object and match directly to functions in its namespace.
We don't need anymore to compile test namespaces and the code is simpler.